decouple analytics code from ga implementation
